fix subdir handling
